requirements.txt
python-telegram-bot>=20.0
google-auth
google-auth-oauthlib
google-api-python-client
google-cloud-aiplatform
pillow
pydantic
python-dotenv
apscheduler
requests
pytest
opencv-python
numpy
pandas
matplotlib
PyMuPDF